+++ src/Content/Backend/NV.Templates.Backend.Core/CoreServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using HttpTracing;
+using Microsoft.Extensions.Configuration;
 using NV.Templates.Backend.Core.General;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -11,6 +12,14 @@ namespace Microsoft.Extensions.DependencyInjection
         public static IServiceCollection AddCore(this IServiceCollection services, IConfiguration configuration)
         {
             // services.BindOptionsToConfigurationAndValidate<>(configuration);
+
+            services.AddHttpTracingToAllHttpClients((sp, builder) =>
+            {
+                return new HttpMessageHandlerTracingConfiguration
+                {
+                    BufferRequests = true,
+                };
+            });
 
             services.AutoRegisterServicesFromAssemblyContaining<IApplicationInfo>();